// netshoes-frontend/types/products.ts
export interface ProductDetails {
  name: string;
  description: string;
}

export interface Product {
  code: string;
  name: string;
  available: boolean;
  visible: boolean;
  details: ProductDetails;
  priceInCents: string;
  salePriceInCents: string;
  rating: number;
  image: string;
  stockAvailable: boolean;
}

export interface ProductsData {
  total: number;
  pageSize: number;
  totalPages: number;
  products: Product[];
}